refactor: 미사용 클래스 삭제 및 패키지 변경
