+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/core/ClubQueryDslRepository.java
@@ -1,0 +1,8 @@
+package com.mobble.mobbleserver.infrastructure.persistence.club.core;
+
+import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.ClubLikeInfoDto;
+
+public interface ClubQueryRepository {
+
+    ClubLikeInfoDto findLikeInfoByClubIdAndMemberId(Long clubId, Long memberId);
+}
