+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/core/projection/ClubLikeInfoDto.java
