--- src/main/java/com/mobble/mobbleserver/refactor/club/core/repository/ClubQueryRepository.java
@@ -1,8 +1,0 @@
-package com.mobble.mobbleserver.refactor.club.core.repository;
-
-import com.mobble.mobbleserver.refactor.club.core.repository.dto.ClubLikeInfoDto;
-
-public interface ClubQueryRepository {
-
-    ClubLikeInfoDto findLikeInfoByClubIdAndMemberId(Long clubId, Long memberId);
-}
