+++ src/test/java/com/mobble/mobbleserver/infrastructure/persistence/club/core/repository/ClubRepositoryImplTest.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.refactor.club.core.repository;
 
 import com.mobble.mobbleserver.config.QueryDslConfig;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
-import com.mobble.mobbleserver.refactor.club.core.repository.dto.ClubLikeInfoDto;
+import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.ClubLikeInfoDto;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.like.clubLike.entity.ClubLike;
 import com.mobble.mobbleserver.domain.member.Member;
@@ -22,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ClubRepositoryImplTest {
 
     @Autowired
-    private ClubRepository clubRepository;
+    private ClubDslRepository clubRepository;
 
     @Autowired
     private EntityManager em;
